docserve/pyproject.toml
[project]
name = "docserve"
version = "0.1.0"
description = "🙃"
authors = [
    {name = "denver",email = "denver@semiotic.ai"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.0"
graphdoc = {path = "../graphdoc/graphdoc", develop = true}  
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
